Work around bugs in Apple GCC 4.2.1
FossilOrigin-Name: cbc09c6e268957178078ad677b58a68217833c4f5aacca88d66f6ace180c1ccc

+++ src/OFDate.m
@@ -470,7 +470,8 @@ tmAndTzToTime(struct tm *tm, int16_t *tz)
 {
 	void *pool = objc_autoreleasePoolPush();
 	int64_t seconds = (int64_t)_seconds;
-	uint32_t nanoseconds = (_seconds - trunc(_seconds)) * 1000000000;
+	uint32_t nanoseconds =
+	    (uint32_t)((_seconds - trunc(_seconds)) * 1000000000);
 	OFData *ret;
 
 	if (seconds >= 0 && seconds < 0x400000000) {
