+++ src/platform/posix/condition.m
@@ -50,7 +50,7 @@ of_condition_timed_wait(of_condition_t *condition, of_mutex_t *mutex,
 	struct timespec ts;
 
 	ts.tv_sec = (time_t)timeout;
-	ts.tv_nsec = (timeout - ts.tv_sec) * 1000000000;
+	ts.tv_nsec = (long)((timeout - ts.tv_sec) * 1000000000);
 
 	return (pthread_cond_timedwait(condition, mutex, &ts) == 0);
 }
